requirements.txt
discord.py
Flask
flask-cors
yt-dlp
python-dotenv
gunicorn